requirements.txt
flask==2.0.1
requests==2.26.0
mysql-connector-python==8.0.26 